# crypto/x509,math/rand/v2: implement the encoding.(Binary|Text)Appender