--- a/doc/next/6-stdlib/99-minor/crypto/x509/62384.md
+++ b/doc/next/6-stdlib/99-minor/crypto/x509/62384.md
@@ -1,0 +1,2 @@
+[OID] now implements the [encoding.BinaryAppender] and [encoding.TextAppender]
+interfaces.