--- a/doc/next/6-stdlib/99-minor/math/rand/v2/62384.md
+++ b/doc/next/6-stdlib/99-minor/math/rand/v2/62384.md
@@ -1,0 +1,1 @@
+[ChaCha8] and [PCG] now implement the [encoding.BinaryAppender] interface.